Changed types for more accurate input

+++ src/main/java/rrr/core/models/entities/Receipt.java
@@ -26,11 +26,11 @@ public class Receipt {
 
     private String date;
 
-    private int tax;
+    private float tax;
 
-    private int tip;
+    private float tip;
 
-    private int total;
+    private float total;
 
     private String description;
 
@@ -54,27 +54,27 @@ public class Receipt {
         this.date = date;
     }
 
-    public int getTax() {
+    public float getTax() {
         return tax;
     }
 
-    public void setTax(int tax) {
+    public void setTax(float tax) {
         this.tax = tax;
     }
 
-    public int getTip() {
+    public float getTip() {
         return tip;
     }
 
-    public void setTip(int tip) {
+    public void setTip(float tip) {
         this.tip = tip;
     }
 
-    public int getTotal() {
+    public float getTotal() {
         return total;
     }
 
-    public void setTotal(int total) {
+    public void setTotal(float total) {
         this.total = total;
     }
